requirements.txt
asgiref==3.5.1
attrs==21.4.0
coverage==6.3.2
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.5.1
inflection==0.5.1
jsonschema==4.5.1
PyJWT==2.3.0
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
